# frontend/src/types/openapi.json
{
  "components": {
    "schemas": {
...
      "BrowserSessionCreate": {
        "description": "Create/update a persistent browser session profile.",
        "properties": {
          "auth_config_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth Config Id"
          },
          "auto_bind_sources": {
            "default": true,
            "title": "Auto Bind Sources",
            "type": "boolean"
          },
          "profile_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Name"
          },
          "site_url": {
            "title": "Site Url",
            "type": "string"
          }
        },
        "required": [
          "site_url"
        ],
        "title": "BrowserSessionCreate",
        "type": "object"
      },
      "BrowserSessionOpenLoginRequest": {
        "description": "Bootstrap persistent browser profile (MVP: optional cookie seeding + page open).\n\nWhen ``headless`` is False, ``dwell_seconds`` is the **upper bound**: the\nserver waits until the user closes the browser window (capturing the\nlogin state), or until the timeout fires. Bump the default/upper limit\nfor manual-login flows (NYT, WSJ, …) that need time for captchas/2FA.",
        "properties": {
          "bootstrap_auth_cookies": {
            "default": true,
            "title": "Bootstrap Auth Cookies",
            "type": "boolean"
          },
          "dwell_seconds": {
            "default": 180,
            "maximum": 900.0,
            "minimum": 0.0,
            "title": "Dwell Seconds",
            "type": "integer"
          },
          "headless": {
            "default": true,
            "title": "Headless",
            "type": "boolean"
          },
          "sync_cookies_to_auth_config": {
            "default": true,
            "title": "Sync Cookies To Auth Config",
            "type": "boolean"
          }
        },
        "title": "BrowserSessionOpenLoginRequest",
        "type": "object"
      },
      "BrowserSessionValidateRequest": {
        "description": "Validate whether session can access first-party article page.",
        "properties": {
          "min_article_paragraphs": {
            "default": 3,
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Min Article Paragraphs",
            "type": "integer"
          },
          "sync_cookies_to_auth_config": {
            "default": true,
            "title": "Sync Cookies To Auth Config",
            "type": "boolean"
          },
          "test_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Test Url"
          },
          "wait_ms": {
            "default": 3500,
            "maximum": 30000.0,
            "minimum": 0.0,
            "title": "Wait Ms",
            "type": "integer"
          }
        },
        "title": "BrowserSessionValidateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "个人化资讯监控管理系统 API",
    "title": "Personal Information Monitor",
    "version": "1.4.3"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/configs/browser-sessions": {
      "get": {
        "operationId": "list_browser_sessions_api_configs_browser_sessions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Browser Sessions",
        "tags": [
          "configs"
        ]
      },
      "post": {
        "operationId": "create_browser_session_api_configs_browser_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrowserSessionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Create Browser Session",
        "tags": [
          "configs"
        ]
      }
    },
    "/api/configs/browser-sessions/{session_id}": {
      "delete": {
        "description": "Remove a persistent browser session.\n\nThe on-disk user-data directory is intentionally kept so that a user can\nrecover the login state by re-creating a session with the same site_url.\nIf they want to wipe local state, they can delete the ``user_data_dir``\npath manually (returned by the list endpoint).",
        "operationId": "delete_browser_session_api_configs_browser_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Delete Browser Session",
        "tags": [
          "configs"
        ]
      }
    },
    "/api/configs/browser-sessions/{session_id}/bind-sources": {
      "post": {
        "description": "Bind every matching monitoring source to this browser session.\n\nBrowser sessions are naturally one-to-many (a single x.com login covers\nevery X feed; a single nytimes.com login covers every NYT RSS/article\nsource). This endpoint exposes\n:func:`bind_browser_session_to_sources` as an explicit user action so\nusers can rebind after adding new sources or after losing the link via\nlegacy cleanup, without having to re-create the session. For X sessions\nmissing an ``auth_config_id`` (e.g. the shared row was deleted earlier),\nwe also self-heal the link before binding, otherwise the X branch of the\nbinder has nothing to point sources at.",
        "operationId": "rebind_browser_session_sources_api_configs_browser_sessions__session_id__bind_sources_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Rebind Browser Session Sources",
        "tags": [
          "configs"
        ]
      }
    },
    "/api/configs/browser-sessions/{session_id}/open-login": {
      "post": {
        "operationId": "open_browser_session_login_api_configs_browser_sessions__session_id__open_login_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrowserSessionOpenLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Open Browser Session Login",
        "tags": [
          "configs"
        ]
      }
    },
    "/api/configs/browser-sessions/{session_id}/validate": {
      "post": {
        "operationId": "validate_browser_session_api_configs_browser_sessions__session_id__validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrowserSessionValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Validate Browser Session",
        "tags": [
          "configs"
        ]
      }
    },
...
  }
}